HomeView refactor
